[#7794] YSQL: Add option to forward pggate RPCs to local tserver
Each postgres process connects to all the Tservers in the cluster which
leads to a quadratic growth in the number of connections in the cluster.
To reduce the number of connections, we can forward all the rpcs to the
local Tserver which can then forward the rpc to the appropriate Tserver.

This feature involves 2 changes:

Forward all the rpcs from pggate to the local tserver.
Ensure the local tserver sends the rpcs to the appropriate tserver.
First change involves creating a RemoteTabletServer for the local tserver
and using that for all the PGSQL read/write rpcs. This needs the postgres
process to be passed the UUID of the tserver process so that it can identify
the HostPort of the local tserver from the list of all tservers.

The second change involves creating the ForwardRpc class which forwards
the read/write request protobufs to the appropriate tserver by using the
TabletInvoker. This class is invoked by the TabletServiceImpl on realizing that
the request is not meant for itself.

This feature is controlled by the GFLAG: ysql_forward_rpcs_to_local_tserver
which allows us to completely enable/disable the forwarding of rpcs to the
local tserver.

Test Plan: Java tests for pggate

Reviewers:
amitanand, bogdan, mihnea, timur, dmitry

Subscribers:
yql, ybase