[WIP][YSQL] Add an option to disable transactions for writes.
The writes in our system perform a distributed transaction if they are
not a single row write. These writes are slow due to the extra
round trips involved for persisting the transaction metadata.

For certain operations like the initial load of data in TPCC, or a copy
operation, we can use the non-transactional code path for speeding up
the writes.

Reviewers:
mihnea

Subscribers:
yql